+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.1
+
+## Added
+- book content - online read function မှာ -> download လုပ်ထားတဲ့ book ရှိနေရင် auto ဖတ်လို့ရသွားမယ်။ 
+## Fixed
+- pdf reader fullscreen error
+
 ## 1.0.0
 
 - init release version